requirements.txt
absl-py==0.15.0
astunparse==1.6.3
backcall==0.2.0
bidict==0.21.4
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
clang==5.0
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
flatbuffers==1.12
gast==0.4.0
google-auth==2.3.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
h5py==3.1.0
idna==3.3
-e git+https://github.com/jordiyapz/infant.git@0956e803ee850c0a972ca2590a42767af6aff29f#egg=infant
ipykernel==6.4.2
ipython==7.29.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==7.0.6
jupyter-core==4.9.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.4.58
opt-einsum==3.3.0
pandas==1.3.4
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.21
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
pyparsing==3.0.4
python-dateutil==2.8.2
python-engineio==4.3.0
python-socketio==5.4.1
pytz==2021.3
pywin32==302
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.7.0
termcolor==1.1.0
torch==1.10.0
torchvision==0.11.1
tornado==6.1
traitlets==5.1.1
typing-extensions==3.7.4.3
urllib3==1.26.7
wcwidth==0.2.5
websocket-client==1.2.1
Werkzeug==2.0.2
wrapt==1.12.1
